fix: Final fix for regressions and build errors in HomeFragment
This commit addresses all remaining issues after the major refactoring of the home screen to an MVVM architecture.

- **Fix Memory Leak:** A critical memory leak in `HeaderAdapter.kt` has been resolved by correctly removing the Firebase `ValueEventListener` in the `onViewRecycled` method.
- **Fix Build Errors:** All `Long` vs `Int` type mismatch build errors have been permanently fixed by explicitly casting hexadecimal color literals to `Int` using `.toInt()` in both `HeaderAdapter.kt` and `PostAdapter.kt`.
- **Restore UI Functionality:** Functional regressions, including the immediate update of the like count and the correct avatar cache lookup, have been fixed in `PostAdapter.kt`.

This completes the refactoring of `HomeFragment`, resulting in a stable, modern, and maintainable codebase.

+++ app/src/main/java/com/synapse/social/studioasinc/home/PostAdapter.kt
@@ -90,8 +90,8 @@ class PostAdapter(
                 cornerRadius = 300f
                 setColor(Color.TRANSPARENT)
             }
-            _imageColor(postPrivateStateIcon, 0xFF616161)
-            _viewGraphics(topMoreButton, 0xFFFFFFFF, 0xFFEEEEEE, 300.0, 0.0, Color.TRANSPARENT)
+            _imageColor(postPrivateStateIcon, 0xFF616161.toInt())
+            _viewGraphics(topMoreButton, 0xFFFFFFFF.toInt(), 0xFFEEEEEE.toInt(), 300.0, 0.0, Color.TRANSPARENT)
 
             post.post_text?.let {
                 MarkdownRenderer.get(context).render(postMessageTextMiddle, it)
